Renaming fixes. Limit baseline percentage by 100%.

+++ contracts/reward/AccumulatingRewardPool.sol
@@ -17,7 +17,7 @@ abstract contract AccumulatingRewardPool is BasicRewardPool {
   constructor(
     IRewardController controller,
     uint256 initialRate,
-    uint256 baselinePercentage
+    uint16 baselinePercentage
   ) public BasicRewardPool(controller, initialRate, baselinePercentage) {}
 
   function isLazy() public view override returns (bool) {
